feat: 내가 푼 문제 개수, 내가 제출한 질문 개수 구하는 로직 추가

+++ backend/src/main/java/com/backend/api/question/service/QuestionService.java
@@ -144,7 +144,7 @@ public class QuestionService {
     }
 
     @Transactional(readOnly = true)
-    public long countByUser(User user) {
+    public int countByUser(User user) {
         validateUserAuthority(user);
         return questionRepository.CountByAuthor(user);
     }
